refactor(sanity): replace getSettings query
