+++ src/app/(website)/layout.tsx
@@ -7,8 +7,8 @@ import Footer from '@/components/Footer';
 import Header from '@/components/Header';
 import PreviewBanner from '@/components/PreviewBanner';
 import Providers from '@/components/Providers';
-import { getSettings } from '@/lib/api.server';
 import fontVariables from '@/lib/font-variables';
+import { getSettings } from '@/lib/sanity/getSettings';
 import { getMetadata } from '@/lib/seo';
 import clsx from 'clsx';
 import { draftMode } from 'next/headers';
